% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmfit.R
\name{dlmfit}
\alias{dlmfit}
\title{Fit univariate Dynamic Linear Models}
\usage{
dlmfit(y, index = NULL, xreg = NULL, order, freq,
  seasonality = "none", delta, level = 0.05, prior = NULL)
}
\arguments{
\item{y}{a numeric vector or a \code{ts} class object specify the time series.}

\item{index}{Optionally, a vector for the time index variable.}

\item{xreg}{Optionally, a vector or matrix of external regressors, which must have the same number of rows as x.}

\item{order}{order of the polynomial model. The default corresponds to second order model.}

\item{freq}{number of seasons to create a DLM with seasonal factors or period to create Fourier representation of a periodic DLM component.}

\item{seasonality}{could be \code{none}, \code{factor} or \code{fourier}.}

\item{delta}{a vector indicating discount factors \eqn{0< \delta_i \leq 1} for \eqn{i=1,\ldots,k}
components of DLM.}

\item{level}{cumulative probability (\eqn{\alpha}) to compute confidence limits. The default is \code{qn = 0.05}.}

\item{prior}{a list to specify the prior parameters \eqn{\bold{m}_0},  \eqn{\bold{C}_0}, \eqn{n_0} and \eqn{S_0}.}
}
\value{
A tidy \code{tsibble} with the columns:
\describe{
 \item{index}{time index variable}
 \item{value}{the observed value at time t}
 \item{theta_k}{the posterior mean at time \eqn{t} for the \code{k}th component of \eqn{\bold{\theta}_t}}
 \item{dC_k}{the posterior variance at time \eqn{t} for the \code{k}th component of \eqn{\bold{\theta}_t}}
 \item{\code{Lw_k}; \code{Up_k}}{the \eqn{\alpha/2} and \eqn{1-\alpha/2} quantiles, respectively, of
   the posterior distribution at time \eqn{t} for the \code{k}th component of \eqn{\bold{\theta}_t}}
 \item{\code{f}}{the mean of the one-step ahead predictive distribution at time \eqn{t}}
 \item{\code{Q}}{the variance of the one-step ahead predictive distribution at time \eqn{t}}
 \item{\code{Lw}; \code{Up}}{the  \eqn{\alpha/2} and \eqn{1-\alpha/2}
   quantiles, respectively, of the predictive distribution at time \eqn{t}.}
}
}
\description{
The function fits univariate Dynamic Linear Models with polynomial, season and/or harmonic
structures by using discount factors for \eqn{\bold{W}_t}.
}
\details{
Inference procedure is implemented following the unified theoretical
of the univariate DLM. For further details see West and Harrison (1997).
}
\examples{
y <- AirPassengers
fit_poly <- dlmfit(y = y, order = 2, delta = 0.98)
fit_poly

fit_seas1 <- dlmfit(y = y, order = 2, freq = 3, seasonality = "fourier", delta = c(0.98, 0.95))
fit_seas1

fit_seas2 <- dlmfit(y = y, order = 2, freq = 12, seasonality = "factor", delta = c(0.90, 0.95))
fit_seas2


}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
