% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/smooth.R
\name{kalmanDist}
\alias{kalmanDist}
\alias{filterDist}
\alias{smoothDist}
\title{Fitting Dynamic Models}
\usage{
filterDist(
  model,
  y,
  prior = NULL,
  var_law = "identity",
  pow = 1,
  delta_phi = 1
)

smoothDist(
  model,
  y,
  prior = NULL,
  var_law = "identity",
  pow = 1,
  delta_phi = 1
)
}
\arguments{
\item{model}{\code{dm} class object.}

\item{y}{observed data.}

\item{prior}{list specifying the prior parameters.}

\item{var_law}{name of the variance law.}

\item{pow}{parameter power for the variance law.}

\item{delta_phi}{discount factor for variance evolution.}
}
\description{
Fitting univariate Dynamic Models via discount factor principle.

Get the smooth distribution for univariate Dynamic Models.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.

West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}

André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
