% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\alias{StatePoly}
\alias{poly.dm}
\alias{seas.dm}
\alias{trig.dm}
\alias{superposition.dm}
\alias{StateSeas}
\alias{StateTrig}
\alias{superposition}
\title{Dynamic models}
\usage{
StatePoly(order, delta = 0.98)

StateSeas(frequency, delta = 0.96)

StateTrig(s, delta = 0.96)

superposition(..., lt)
}
\arguments{
\item{order}{order of the polynomial model.}

\item{delta}{discount factor.}

\item{frequency}{number of seasons to create seasonal factors representation.}

\item{s}{the period to create Fourier representation.}

\item{...}{pass the models separate by commma to make the superposition.}

\item{lt}{list of models to make the superposition.}
}
\description{
Available Dynamic Models.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
