% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdlm.R
\name{plotdlm}
\alias{plotdlm}
\title{Plots for fitted DLM from \code{dlmfit} function}
\usage{
plotdlm(mod, which = c("level", "thetas", "pred", "all"))
}
\arguments{
\item{mod}{the output from \code{dlmfit} function, a \code{tsibble} object.}

\item{which}{a character indicating the type of plot. The options are
(i) \code{level} for the posterior mean level with confidence limits and the observed data,
(ii) \code{pred} for the mean of predictive distribution with confidence limits and the observed data,
(iii) \code{theta} for the mean posterior paramters with confidence limits and
(vi) \code{all} for all plots.}
}
\value{
The plot(s) according to \code{which} option.
}
\description{
The function provides several plots of posteriori and predictive distributions.
}
\examples{

y <- AirPassengers
fit_poly = dlmfit(y = y, order = 2, delta = 0.95)
plotdlm(fit_poly, which="level")
plotdlm(fit_poly, which="pred")
plotdlm(fit_poly, which="thetas")

fit_seas <- dlmfit(y = y, order = 2, freq = 3, seasonality = "fourier", delta = c(0.95, 0.95))
plotdlm(fit_seas, which="level")
plotdlm(fit_seas, which="pred")
plotdlm(fit_seas, which="thetas")

}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\seealso{
\code{\link{dlmfit}}
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
