% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\alias{plot.dm}
\alias{PlotPredictive}
\alias{PlotState}
\alias{PlotForecast}
\alias{PlotSmooth}
\alias{PlotResiduals}
\title{Plottting functions}
\usage{
\method{plot}{dm}(x, ...)

PlotPredictive(
  object,
  interval = TRUE,
  geom_data = "line",
  interval.type = "exact",
  interval.level = 0.05
)

PlotState(
  object,
  which.state = "all",
  interval.type = "exact",
  interval.level = 0.05
)

PlotForecast(
  object,
  horizon,
  which.distr = "predictive",
  which.state = "all",
  interval.type = "exact",
  interval.level = 0.05
)

PlotResiduals(object)

PlotSmooth(
  object,
  which.distr = "mean",
  which.state = "all",
  interval.type = "exact",
  interval.level = 0.05
)
}
\arguments{
\item{x}{não sei}

\item{...}{currently not used.}

\item{object}{\code{dm} class object from \code{fit} or \code{smooth} functions.}

\item{interval}{logical indicating if the interval will be plot}

\item{geom_data}{if the observed data will be represent by point or lines}

\item{interval.type}{which type of interval to use}

\item{interval.level}{which significance level to use}

\item{which.state}{which state to plot}

\item{horizon}{horizon of forecast}

\item{which.distr}{which distribution to plot}
}
\description{
Functions to plot the results of a model fit or smooth.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
