% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdlmPoly.R
\name{rdlmPoly}
\alias{rdlmPoly}
\title{Simulate univariate DLM with polynomial structure}
\usage{
rdlmPoly(n, order, m0, C0, V, delta = 0.95)
}
\arguments{
\item{n}{number of observations.}

\item{order}{order of the polynomial model.}

\item{m0}{vector for prior mean (given \eqn{D_0}).}

\item{C0}{matrix for prior variance (given \eqn{D_0}).}

\item{V}{numeric value for the observational variance \eqn{V}.}

\item{delta}{discount factor \eqn{0< \delta \leq 1} to compute the evolution variance matrices \eqn{\bold{W}_t}.}
}
\value{
A list with two entries: \code{y} the simulated series and \code{theta} the simulated parameters.
}
\description{
The function simulate values (\eqn{y} and \eqn{\bold{\theta}})
from a univariate DLM with polynomial structure.
}
\details{
To simulate we use the hierarchical represention of DLM. For further details see the package vignette.
}
\examples{

set.seed(1212)
out = rdlmPoly(n = 100, order = 1, m0 = 2, C0 = 0.5, V=1.5, delta=0.9)
plot(out[['y']], type = 'l')
plot(out[['theta']][, 1], type = 'l')

set.seed(1212)
out = rdlmPoly(n = 100, order = 2, m0 = c(1.0, 0.2), C0 = diag(0.5, 2), V=1.5, delta=0.9)
plot(out[['y']], type = 'l')
plot(out[['theta']][, 1], type = 'l')
plot(out[['theta']][, 2], type = 'l')



}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\seealso{
\code{\link{dlmfit}}
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
