% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_moments.R
\name{update_moments}
\alias{update_moments}
\title{Compute the posterior and predictive moments at time \eqn{t}.}
\usage{
update_moments(y, t, FF, GG, R, m0, C0, S0, n0, level)
}
\arguments{
\item{y}{observed value at time \eqn{t}.}

\item{t}{observed time.}

\item{FF}{vector \eqn{\bold{FF}_t} of the model.}

\item{GG}{matrix \eqn{\bold{GG}_t} of the model.}

\item{R}{matrix \eqn{\bold{R} = \bold{P} + \bold{W}} of the model.}

\item{m0}{prior mean at time \eqn{t} for \eqn{\bold{\theta}_t}.}

\item{C0}{prior variance at time \eqn{t} for \eqn{\bold{\theta}_t}.}

\item{S0}{prior variance at time \eqn{t} for \eqn{\phi=V^{-1}}.}

\item{n0}{prior shape parameter of the Gamma distribuition at time \eqn{t} for \eqn{\phi=V^{-1}}.}

\item{level}{cumulative probability (\eqn{\alpha}) to compute confidence limits.}
}
\description{
The function is used in \code{dlmfit} to calculate recursively the expected value and variance of posterior
and predictive distributions at time \eqn{t}.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
